+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Added minimum priority optional argument to "standup" command
 - Added numbering to standup
+- Added "yesterday" command to mark completed tasks you forgot about
 
 ### Fixed
 - Indexing fails when no addon directory set